renamed package
 On branch george/addWorkAllocatorNext
 Changes to be committed:
	renamed:    src/main/java/me/itzg/tryetcdworkpart/Bits.java -> src/main/java/com/rackspace/salus/common/workpart/Bits.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/Work.java -> src/main/java/com/rackspace/salus/common/workpart/Work.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/TryEtcdWorkPartitionerApplication.java -> src/main/java/com/rackspace/salus/common/workpart/WorkPartitionerApplication.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/WorkProcessor.java -> src/main/java/com/rackspace/salus/common/workpart/WorkProcessor.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/config/EtcdConfig.java -> src/main/java/com/rackspace/salus/common/workpart/config/EtcdConfig.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/config/EtcdProperties.java -> src/main/java/com/rackspace/salus/common/workpart/config/EtcdProperties.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/config/WorkerProperties.java -> src/main/java/com/rackspace/salus/common/workpart/config/WorkerProperties.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/lombok.config -> src/main/java/com/rackspace/salus/common/workpart/lombok.config
	renamed:    src/main/java/me/itzg/tryetcdworkpart/services/DefaultWorkProcessor.java -> src/main/java/com/rackspace/salus/common/workpart/services/DefaultWorkProcessor.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/services/WorkAllocator.java -> src/main/java/com/rackspace/salus/common/workpart/services/WorkAllocator.java
	renamed:    src/main/java/me/itzg/tryetcdworkpart/web/WorkController.java -> src/main/java/com/rackspace/salus/common/workpart/web/WorkController.java
	renamed:    src/test/java/me/itzg/tryetcdworkpart/services/WorkAllocatorTest.java -> src/test/java/com/rackspace/salus/common/workpart/services/WorkAllocatorTest.java
