+++ src/main/java/com/rackspace/salus/common/workpart/WorkPartitionerApplication.java
@@ -1,12 +1,12 @@
-package me.itzg.tryetcdworkpart;
+package com.rackspace.salus.common.workpart;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TryEtcdWorkPartitionerApplication {
+public class WorkPartitionerApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(TryEtcdWorkPartitionerApplication.class, args);
+    SpringApplication.run(WorkPartitionerApplication.class, args);
   }
 }
