+++ src/main/java/com/rackspace/salus/common/workpart/WorkProcessor.java
@@ -1,4 +1,4 @@
-package me.itzg.tryetcdworkpart;
+package com.rackspace.salus.common.workpart;
 
 public interface WorkProcessor {
