+++ src/main/java/com/rackspace/salus/common/workpart/config/EtcdProperties.java
@@ -1,4 +1,4 @@
-package me.itzg.tryetcdworkpart.config;
+package com.rackspace.salus.common.workpart.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
