+++ src/main/java/com/rackspace/salus/common/workpart/services/WorkAllocator.java
@@ -1,17 +1,17 @@
-package me.itzg.tryetcdworkpart.services;
+package com.rackspace.salus.common.workpart.services;
 
 import static com.coreos.jetcd.data.ByteSequence.fromString;
 import static com.coreos.jetcd.op.Op.delete;
 import static com.coreos.jetcd.op.Op.put;
-import static me.itzg.tryetcdworkpart.Bits.ACTIVE_SET;
-import static me.itzg.tryetcdworkpart.Bits.REGISTRY_SET;
-import static me.itzg.tryetcdworkpart.Bits.WORKERS_SET;
-import static me.itzg.tryetcdworkpart.Bits.WORK_LOAD_FORMAT;
-import static me.itzg.tryetcdworkpart.Bits.extractIdFromKey;
-import static me.itzg.tryetcdworkpart.Bits.fromFormat;
-import static me.itzg.tryetcdworkpart.Bits.isDeleteKeyEvent;
-import static me.itzg.tryetcdworkpart.Bits.isNewKeyEvent;
-import static me.itzg.tryetcdworkpart.Bits.isUpdateKeyEvent;
+import static com.rackspace.salus.common.workpart.Bits.ACTIVE_SET;
+import static com.rackspace.salus.common.workpart.Bits.REGISTRY_SET;
+import static com.rackspace.salus.common.workpart.Bits.WORKERS_SET;
+import static com.rackspace.salus.common.workpart.Bits.WORK_LOAD_FORMAT;
+import static com.rackspace.salus.common.workpart.Bits.extractIdFromKey;
+import static com.rackspace.salus.common.workpart.Bits.fromFormat;
+import static com.rackspace.salus.common.workpart.Bits.isDeleteKeyEvent;
+import static com.rackspace.salus.common.workpart.Bits.isNewKeyEvent;
+import static com.rackspace.salus.common.workpart.Bits.isUpdateKeyEvent;
 
 import com.coreos.jetcd.Client;
 import com.coreos.jetcd.Lease.KeepAliveListener;
@@ -42,10 +42,10 @@ import java.util.concurrent.Semaphore;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 import lombok.extern.slf4j.Slf4j;
-import me.itzg.tryetcdworkpart.Bits;
-import me.itzg.tryetcdworkpart.Work;
-import me.itzg.tryetcdworkpart.WorkProcessor;
-import me.itzg.tryetcdworkpart.config.WorkerProperties;
+import com.rackspace.salus.common.workpart.Bits;
+import com.rackspace.salus.common.workpart.Work;
+import com.rackspace.salus.common.workpart.WorkProcessor;
+import com.rackspace.salus.common.workpart.config.WorkerProperties;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.SmartLifecycle;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
